requirements.txt
pandas>=1.5
numpy
scikit-learn>=1.3
requests
joblib
tqdm
python-dateutil
python-dotenv
nba_api>=1.2.1
